third_party/WebKit/LayoutTests/editing/execCommand/insert-list-with-progress-crash.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(()=> assert_selection(
  [
  '<div contenteditable>',
    '<progress hidden>',
      '<textarea></textarea>',
     '</progress>',
    '^<select></select>|',
  '<div>'],
  'InsertOrderedList',
  [
  '<div contenteditable>',
    '<ol>',
      '<li>',
        '^<select></select>|',
      '</li>',
    '</ol>',
    '<progress hidden>',
      '<textarea></textarea>',
    '</progress>',
    '<div></div>',
    '</div>']),
  "This InsertOrderedList in contenteditable div with progress.");
</script>
